requirements.txt
alabaster==0.7.16
babel==2.16.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudscraper==1.2.71
dnspython==1.16.0
docutils==0.20.1
feedparser==6.0.11
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
gnews==0.3.7
idna==3.10
imagesize==1.4.1
indic-nlp-library==0.92
itsdangerous==2.2.0
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
lxml==5.3.0
lxml_html_clean==0.2.2
MarkupSafe==2.1.5
Morfessor==2.0.6
newspaper4k==0.9.3.1
nltk==3.9.1
numpy==1.26.4
packaging==24.1
pandas==2.2.3
pillow==10.4.0
Pygments==2.18.0
pyparsing==3.1.4
pythainlp==5.0.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
sgmllib3k==1.0.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==7.4.7
sphinx-argparse==0.5.2
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
tinysegmenter==0.4
tldextract==5.1.2
tomli==2.0.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
